Cleanup: Split SetProgramRunParams implementation into a separate header
